Former-commit-id: a01e29716f64e31d1f4871f4f23898745ae447a6

+++ Projects/UnitTests/Classes/UIListTest.h
@@ -54,8 +54,8 @@ public:
 	// UIListDelegate
     virtual int32 ElementsCount(UIList *list);
 	virtual UIListCell *CellAtIndex(UIList *list, int32 index);
-	virtual int32 CellWidth(UIList *list, int32 index);
-	virtual int32 CellHeight(UIList *list, int32 index);
+	virtual float32 CellWidth(UIList *list, int32 index);
+	virtual float32 CellHeight(UIList *list, int32 index);
 
 private:
 	Vector2 cellSize;
